requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
pydantic==2.10.3
SQLAlchemy==2.0.36
requests==2.32.3
python-dotenv==1.0.1
yfinance==0.2.54
pandas==2.2.3
numpy==2.1.3
textblob==0.18.0.post0
